IGNITE-24286 Move StringCacheClassTest to StringCacheTest
